backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
python-multipart
python-dotenv
